Refactored the report manager and added missing comment headers
