+++ pstatus-graphql-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/services/ReportMutationService.kt
@@ -5,15 +5,14 @@ import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.google.gson.ToNumberPolicy
 import gov.cdc.ocio.database.models.Report
-import gov.cdc.ocio.database.persistence.ProcessingStatusRepository
 import gov.cdc.ocio.database.utils.DateLongFormatTypeAdapter
 import gov.cdc.ocio.database.utils.InstantTypeAdapter
 import gov.cdc.ocio.database.utils.OffsetDateTimeTypeAdapter
 import gov.cdc.ocio.processingstatusapi.exceptions.BadRequestException
-import gov.cdc.ocio.processingstatusapi.exceptions.BadStateException
 import gov.cdc.ocio.processingstatusapi.exceptions.ContentException
 import gov.cdc.ocio.processingstatusapi.extensions.snakeToCamelCase
 import gov.cdc.ocio.processingstatusapi.collections.BasicHashMap
+import gov.cdc.ocio.processingstatusapi.models.Action
 import gov.cdc.ocio.processingstatusapi.mutations.models.UpsertReportResult
 import gov.cdc.ocio.processingstatusapi.mutations.models.ValidatedReportResult
 import gov.cdc.ocio.processingstatusapi.services.ValidationComponents.gson
@@ -25,13 +24,14 @@ import gov.cdc.ocio.reportschemavalidator.utils.DefaultJsonUtils
 import gov.cdc.ocio.reportschemavalidator.validators.JsonSchemaValidator
 import mu.KLogger
 import mu.KotlinLogging
-import org.koin.core.component.KoinComponent
-import org.koin.core.component.inject
 import java.time.Instant
 import java.time.OffsetDateTime
 import java.util.*
 
 
+/**
+ * Object to warehouse the dependencies for the report schema validation service
+ */
 object ValidationComponents {
     private val objectMapper: ObjectMapper by lazy { ObjectMapper() }
     val jsonUtils: DefaultJsonUtils by lazy { DefaultJsonUtils(objectMapper) }
@@ -66,13 +66,11 @@ object ValidationComponents {
  *   into usable formats.
  *
  */
-class ReportMutationService : KoinComponent {
+class ReportMutationService {
 
     private val logger = KotlinLogging.logger {}
 
-    private val repository by inject<ProcessingStatusRepository>()
-
-    private val reportsCollection = repository.reportsCollection
+    private val reportManager = ReportManager()
 
     /**
      * Upserts a report based on the provided input and action.
@@ -101,13 +99,13 @@ class ReportMutationService : KoinComponent {
             val validatedReport = validationResult.report!!
 
             // Assign the report ids
-            val reportId = generateNewId()
+            val reportId = UUID.randomUUID().toString()
             validatedReport.id = reportId
             validatedReport.reportId = reportId
 
             when (actionType) {
-                Action.CREATE -> createReport(validatedReport)
-                Action.REPLACE -> replaceReport(validatedReport)
+                Action.CREATE -> reportManager.createReport(validatedReport)
+                Action.REPLACE -> reportManager.replaceReport(validatedReport)
             }
 
             return UpsertReportResult(
@@ -137,6 +135,12 @@ class ReportMutationService : KoinComponent {
         }
     }
 
+    /**
+     * Validates the report with the provided input [Map].
+     *
+     * @param input Map<String, Any?>?
+     * @return ValidatedReportResult
+     */
     private fun validateReport(input: Map<String, Any?>?): ValidatedReportResult {
         try {
             logger.info("The report received: $input and will be converted to Json")
@@ -201,117 +205,13 @@ class ReportMutationService : KoinComponent {
     }
 
     /**
-     * Creates a new report in the database.
+     * Validates the provided action for the report disposition, which is to create or replace an existing report.
      *
-     * This function generates a new ID for the report, validates the provided upload ID,
-     * and attempts to create the report in the Cosmos DB container. If the report ID
-     * is provided or if the upload ID is missing, a BadRequestException is thrown.
-     * If there is an error during the creation process, a ContentException is thrown.
-     *
-     * @param report The report object to be created. Must not have an existing ID and must include a valid upload ID.
-     * @return The created Report object, or null if the creation fails.
-     * @throws BadRequestException If the report ID is provided or if the upload ID is missing.
-     * @throws ContentException If there is an error during the report creation process.
+     * @param action [String] - Action to be validated
+     * @return [Action] - Validated action as an enumeration
+     * @throws BadRequestException
      */
-    @Throws(BadRequestException::class, ContentException::class)
-    private fun createReport(report: Report) { // TODO - move to ReportManager.kt
-        logger.info("Creating report for uploadId = ${report.uploadId} with stageName = ${report.stageInfo?.action}")
-        val transformedContent = repository.contentTransformer(report.content as Map<*, *>)
-        report.content = transformedContent
-
-        try {
-            val success =
-                reportsCollection.createItem(
-                    generateNewId(),
-                    report,
-                    Report::class.java,
-                    report.uploadId
-                )
-
-            // Check if successful and throw an exception if not
-            if (!success)
-                throw ContentException("Failed to create report")
-
-        } catch (e: Exception) {
-            logger.error(e.localizedMessage)
-            throw ContentException("Failed to create report: ${e.message}")
-        }
-    }
-
-    /**
-     * Replaces an existing report in the database.
-     *
-     * This function checks if the report ID is provided and validates the upload ID.
-     * It attempts to read the existing report from the Cosmos DB container and,
-     * if found, replaces it with the new report data. If the report ID is missing,
-     * or if the upload ID is not provided, a BadRequestException is thrown. If
-     * the report is not found for replacement, another BadRequestException is thrown.
-     * In case of any error during the database operations, an appropriate exception
-     * will be thrown.
-     *
-     * @param report The report object containing the new data. Must have a valid ID and upload ID.
-     * @return The updated Report object, or null if the replacement fails.
-     * @throws BadRequestException If the report ID is missing, the upload ID is missing, or the report is not found.
-     */
-    @Throws(BadRequestException::class, ContentException::class)
-    private fun replaceReport(report: Report) { //TODO- move to ReportManager.kt and update
-        return try {
-            val uploadId = report.uploadId
-            val stageInfo = report.stageInfo
-
-            // Delete all reports matching the report ID with the same service and action name
-            val cName = repository.reportsCollection.collectionNameForQuery
-            val cVar = repository.reportsCollection.collectionVariable
-            val cPrefix = repository.reportsCollection.collectionVariablePrefix
-            val cElFunc = repository.reportsCollection.collectionElementForQuery
-            val sqlQuery = (
-                    "select * from $cName $cVar "
-                            + "where ${cPrefix}uploadId = '$uploadId' "
-                            + "and ${cPrefix}stageInfo.${cElFunc("service")} = '${stageInfo?.service}' "
-                            + "and ${cPrefix}stageInfo.${cElFunc("action")} = '${stageInfo?.action}'"
-                    )
-            val items = repository.reportsCollection.queryItems(
-                sqlQuery,
-                Report::class.java
-            )
-            if (items.isNotEmpty()) {
-                try {
-                    items.forEach {
-                        reportsCollection.deleteItem(
-                            it.id,
-                            it.uploadId
-                        )
-                    }
-                    logger.info("Removed all reports with stage name = $stageInfo?.stage")
-                } catch (e: Exception) {
-                    throw BadStateException("Issue deleting report: ${e.localizedMessage}")
-                }
-            }
-
-            createReport(report)
-
-        } catch (e: Exception) {
-            logger.error(e.localizedMessage)
-            throw ContentException("Failed to replace report: ${e.message}")
-        }
-    }
-
-    /**
-     * Generates a new unique identifier for a report.
-     *
-     * This method creates a UUID string to be used as a unique ID when creating a new report.
-     *
-     * @return A unique string identifier.
-     */
-    private fun generateNewId(): String {
-        // Generate a new unique ID
-        return UUID.randomUUID().toString()
-    }
-
-    enum class Action {
-        CREATE, REPLACE
-    }
-
+    @Throws(BadRequestException::class)
     private fun validateAction(action: String): Action {
         return when (action) {
             "create" -> {
