--- pstatus-graphql-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/utils/ReportManager.kt
@@ -1,4 +1,0 @@
-package gov.cdc.ocio.processingstatusapi.utils
-
-class ReportManager {
-}
